+++ src/etaoin/api.clj
@@ -189,7 +189,8 @@
     nil resp
     (-> resp :value first second)))
 
-(defmethods get-active-element* [:chrome :phantom :safari]
+(defmethods get-active-element*
+  [:chrome :headless :phantom :safari]
   [driver]
   (with-resp driver :post
     [:session (:session @driver) :element :active]
@@ -233,20 +234,32 @@
     nil resp
     (:value resp)))
 
-(defmethods get-window-handles [:chrome :phantom]
+(defmethods get-window-handles
+  [:chrome :headless :phantom]
   [driver]
   (with-resp driver :get
     [:session (:session @driver) :window_handles]
     nil resp
     (:value resp)))
 
-(defn switch-window
+(defmulti switch-window
   "Switches a browser to another window."
   {:arglists '([driver handle])}
+  dispatch-driver)
+
+(defmethod switch-window
+  :default
   [driver handle]
   (with-resp driver :post
     [:session (:session @driver) :window]
     {:handle handle} _))
+
+(defmethods switch-window
+  [:chrome :headless]
+  [driver handle]
+  (with-resp driver :post
+    [:session (:session @driver) :window]
+    {:name handle} _))
 
 (defmulti close-window
   "Closes the current browser window."
@@ -269,7 +282,8 @@
     [:session (:session @driver) :window :maximize]
     nil _))
 
-(defmethods maximize [:chrome :safari]
+(defmethods maximize
+  [:chrome :headless :safari]
   [driver]
   (let [h (get-window-handle driver)]
     (with-resp driver :post
@@ -626,7 +640,8 @@
   {:arglists '([driver])}
   dispatch-driver)
 
-(defmethods mouse-btn-down [:chrome :phantom :safari]
+(defmethods mouse-btn-down
+  [:chrome :headless :phantom :safari]
   [driver]
   (with-resp driver :post
     [:session (:session @driver) :buttondown]
@@ -637,7 +652,8 @@
   {:arglists '([driver])}
   dispatch-driver)
 
-(defmethods mouse-btn-up [:chrome :phantom :safari]
+(defmethods mouse-btn-up
+  [:chrome :headless :phantom :safari]
   [driver]
   (with-resp driver :post
     [:session (:session @driver) :buttonup]
@@ -649,7 +665,8 @@
   {:arglists '([driver q] [driver x y])}
   dispatch-driver)
 
-(defmethods mouse-move-to [:chrome :phantom :safari]
+(defmethods mouse-move-to
+  [:chrome :headless :phantom :safari]
   ([driver q]
    (with-resp driver :post
      [:session (:session @driver) :moveto]
@@ -723,7 +740,8 @@
 
 (defmulti double-click-el dispatch-driver)
 
-(defmethods double-click-el [:chrome :phantom]
+(defmethods double-click-el
+  [:chrome :headless :phantom]
   [driver el]
   (with-resp driver :post
     [:session (:session @driver) :element el :doubleclick]
@@ -734,10 +752,9 @@
 
   Note:
 
-  the supported browsers are Chrome and Phantom.js. For Firefox and
-  Safari, your may try to simulate it as a `click, wait, click`
-  sequence.
-"
+  the supported browsers are (Headless) Chrome, and Phantom.js.
+  For Firefox and Safari, your may try to simulate it as a `click, wait, click`
+  sequence."
   [driver q]
   (double-click-el driver (query driver q)))
 
@@ -747,7 +764,8 @@
 
 (defmulti get-element-size-el dispatch-driver)
 
-(defmethods get-element-size-el [:chrome :phantom :safari]
+(defmethods get-element-size-el
+  [:chrome :headless :phantom :safari]
   [driver el]
   (with-resp driver :get
     [:session (:session @driver) :element el :size]
@@ -775,7 +793,7 @@
 (defmulti get-element-location-el dispatch-driver)
 
 (defmethods get-element-location-el
-  [:chrome :phantom :safari]
+  [:chrome :headless :phantom :safari]
   [driver el]
   (with-resp driver :get
     [:session (:session @driver) :element el :location]
@@ -951,7 +969,8 @@
 
 (defmulti get-active* dispatch-driver)
 
-(defmethods get-active* [:chrome :phantom :safari]
+(defmethods get-active*
+  [:chrome :headless :phantom :safari]
   [driver]
   (with-resp driver :get
     [:session (:session @driver) :element :active]
@@ -1238,7 +1257,9 @@
 ;; alerts
 ;;
 
-(defmulti get-alert-text dispatch-driver)
+(defmulti get-alert-text
+  "Returns a string of text that appears in alert dialog (if present)."
+  dispatch-driver)
 
 (defmethod get-alert-text :firefox
   [driver]
@@ -1248,7 +1269,8 @@
     resp
     (:value resp)))
 
-(defmethods get-alert-text [:chrome :safari]
+(defmethods get-alert-text
+  [:chrome :headless :safari]
   [driver]
   (with-resp driver :get
     [:session (:session @driver) :alert_text]
@@ -1256,7 +1278,9 @@
     resp
     (:value resp)))
 
-(defmulti dismiss-alert dispatch-driver)
+(defmulti dismiss-alert
+  "Simulates cancelling an alert dialog (pressing cross button)."
+  dispatch-driver)
 
 (defmethod dismiss-alert :firefox
   [driver]
@@ -1264,13 +1288,16 @@
     [:session (:session @driver) :alert :dismiss]
     nil _))
 
-(defmethods dismiss-alert [:chrome :safari]
+(defmethods dismiss-alert
+  [:chrome :headless :safari]
   [driver]
   (with-resp driver :post
     [:session (:session @driver) :dismiss_alert]
     nil _))
 
-(defmulti accept-alert dispatch-driver)
+(defmulti accept-alert
+  "Simulates submitting an alert dialog (pressing OK button)."
+  dispatch-driver)
 
 (defmethod accept-alert :firefox
   [driver]
@@ -1278,7 +1305,8 @@
     [:session (:session @driver) :alert :accept]
     nil _))
 
-(defmethods accept-alert [:chrome :safari]
+(defmethods accept-alert
+  [:chrome :headless :safari]
   [driver]
   (with-resp driver :post
     [:session (:session @driver) :accept_alert]
@@ -1332,27 +1360,43 @@
 (defn driver? [driver type]
   (= (dispatch-driver driver) type))
 
-(defn chrome? [driver]
+(defn chrome?
+  "Returns true if a driver is a Chrome instance."
+  [driver]
   (driver? driver :chrome))
 
-(defn firefox? [driver]
+(defn firefox?
+  "Returns true if a driver is a Firefox instance."
+  [driver]
   (driver? driver :firefox))
 
-(defn phantom? [driver]
+(defn phantom?
+  "Returns true if a driver is a Phantom.js instance."
+  [driver]
   (driver? driver :phantom))
 
-(defn safari? [driver]
+(defn safari?
+  "Returns true if a driver is a Safari instance."
+  [driver]
   (driver? driver :safari))
 
-(defn headless? [driver]
+(defn headless?
+  "Returns true if a driver is a Headless Chrome instance."
+  [driver]
   (driver? driver :headless))
 
-(defn exists? [driver q]
+(defn exists?
+  "Returns true if an element exists on the page.
+
+  Keep in mind it does not validates whether the element is visible,
+  clickable and so on."
+  [driver q]
   (with-http-error
     (get-element-text driver q)
     true))
 
-(def absent? (complement exists?))
+(def ^{:doc "Oppsite to `exists?`."}
+  absent? (complement exists?))
 
 (defmulti displayed-el?
   "Checks whether an element is displayed by its identifier.
@@ -1361,8 +1405,7 @@
   have to check some common cases manually (CSS display, visibility,
   etc).
 
-  Returns true or false.
-"
+  Returns true or false."
   dispatch-driver)
 
 (defmethod displayed-el? :default
@@ -1389,11 +1432,14 @@
   [driver q]
   (displayed-el? driver (query driver q)))
 
-(defn visible? [driver q]
+(defn visible?
+  "Checks whether an element is visible on the page."
+  [driver q]
   (and (exists? driver q)
        (displayed? driver q)))
 
-(def invisible? (complement visible?))
+(def ^{:doc "Oppsite to `visible?`."}
+  invisible? (complement visible?))
 
 (defn enabled-el? [driver el]
   (with-resp driver :get
@@ -1410,6 +1456,10 @@
 (def disabled? (complement enabled?))
 
 (defn has-text?
+  "Returns true if text appears anywhere on a page.
+
+  When a query expression is passed, tries to find that text
+  into the first element found with that term."
   ([driver text]
    (has-text? driver {:tag :*} text))
   ([driver q text]
@@ -1431,24 +1481,31 @@
         true)
       false))))
 
-(defn has-class-el? [driver el class]
+(defn has-class-el?
+  [driver el class]
   (let [classes (get-element-attr-el driver el "class")]
     (cond
       (nil? classes) false
       (string? classes)
       (str/includes? classes (name class)))))
 
-(defn has-class? [driver q class]
+(defn has-class?
+  "Checks whether an element has a specific class."
+  [driver q class]
   (has-class-el? driver (query driver q) class))
 
-(def has-no-class? (complement has-class?))
+(def ^{:doc "Opposite to `has-class?`."}
+  has-no-class? (complement has-class?))
 
-(defn has-alert? [driver]
+(defn has-alert?
+  "Checks if there is an alert dialog opened on the page."
+  [driver]
   (with-http-error
     (get-alert-text driver)
     true))
 
-(def has-no-alert? (complement has-alert?))
+(def ^{:doc "Opposite to `has-alert?`."}
+  has-no-alert? (complement has-alert?))
 
 ;;
 ;; wait functions
@@ -1463,6 +1520,15 @@
    (wait sec))
   ([sec]
    (Thread/sleep (* sec 1000))))
+
+(defmacro with-wait
+  "Executes the body waiting for n seconds before each form.
+  Returns a value of the last form. Use that macros to perform
+  a bunch of actions slowly. Some SPA applications need extra time
+  to re-render the content."
+  [n & body]
+  (let [combo (interleave (repeat `(wait ~n)) body)]
+    `(do ~@combo)))
 
 (defn wait-predicate
   "Sleeps continuously calling a predicate until it returns true.
@@ -1644,7 +1710,8 @@
 
 (defmulti touch-tap dispatch-driver)
 
-(defmethod touch-tap :chrome
+(defmethods touch-tap
+  [:chrome :headless]
   [driver q]
   (with-resp driver :post
     [:session (:session @driver) :touch :click]
@@ -1652,7 +1719,8 @@
 
 (defmulti touch-down dispatch-driver)
 
-(defmethod touch-down :chrome
+(defmethods touch-down
+  [:chrome :headless]
   ([driver q]
    (let [{:keys [x y]}
          (get-element-location driver q)]
@@ -1664,7 +1732,8 @@
 
 (defmulti touch-up dispatch-driver)
 
-(defmethod touch-up :chrome
+(defmethod touch-up
+  [:chrome :headless]
   ([driver q]
    (let [{:keys [x y]}
          (get-element-location driver q)]
@@ -1676,7 +1745,8 @@
 
 (defmulti touch-move dispatch-driver)
 
-(defmethod touch-move :chrome
+(defmethod touch-move
+  [:chrome :headless]
   ([driver q]
    (let [{:keys [x y]}
          (get-element-location driver q)]
@@ -1762,24 +1832,24 @@
 ;; input
 ;;
 
-(defn- join-str
-  [text more]
-  (apply str text more))
+(defn- make-input* [text & more]
+  (mapv str (apply str text more)))
 
 (defmulti fill-active*
-  {:arglists '([driver keys])}
+  {:arglists '([driver text & more])}
   dispatch-driver)
 
-(defmethod fill-active* :chrome
-  [driver keys]
+(defmethods fill-active*
+  [:chrome :headless]
+  [driver text & more]
   (with-resp driver :post
     [:session (:session @driver) :keys]
-    {:value (vec keys)} _))
+    {:value (apply make-input* text more)} _))
 
 (defn fill-active
   "Fills an active element with keys."
   [driver text & more]
-  (fill-active* driver (join-str text more)))
+  (apply fill-active* driver text more))
 
 (defmulti fill-el dispatch-driver)
 
@@ -1791,6 +1861,7 @@
     (with-resp driver :post
       [:session (:session @driver) :element el :value]
       {:value (vec keys)} _)))
+
 
 (defmethod fill-el :firefox
   [driver el text]
@@ -1810,7 +1881,7 @@
   (fill driver :simple-input \"foo\" \"baz\" 1)
   ;; fills the input with  \"foobaz1\""
   [driver q text & more]
-  (fill-el driver (query driver q) (join-str text more)))
+  (apply fill-el driver (query driver q) text more))
 
 (defn fill-multi
   "Fills multiple inputs in batch.
@@ -1876,6 +1947,38 @@
   [driver q & more]
   (doseq [q (cons q more)]
     (clear-el driver (query driver q))))
+
+;;
+;; file upload
+;;
+
+(defmulti upload-file
+  "Attaches a local file to a file input field.
+
+  Arguments:
+
+  - `q` is a query term that refers to a file input;
+  - `file` is either a string or java.io.File object
+  that references a local file. The file should exist.
+
+  Under the hood, it sends the file's name as a sequence of keys
+  to the input."
+  (fn [driver q file]
+    (type file)))
+
+(defmethod upload-file String
+  [driver q path]
+  (upload-file driver q (io/file path)))
+
+(defmethod upload-file java.io.File
+  [driver q file]
+  (let [path (.getAbsolutePath file)
+        message (format "File %s does not exist" path)]
+    (if (.exists file)
+      (fill driver q path)
+      (throw+ {:type :etaoin/file
+               :message message
+               :driver @driver}))))
 
 ;;
 ;; submit
@@ -1997,10 +2100,14 @@
   {:arglists '([driver])}
   dispatch-driver)
 
-(defmethods port-args [:firefox :safari] [driver]
+(defmethods port-args
+  [:firefox :safari]
+  [driver]
   ["--port" (:port @driver)])
 
-(defmethods port-args [:chrome :headless] [driver]
+(defmethods port-args
+  [:chrome :headless]
+  [driver]
   [(str "--port=" (:port @driver))])
 
 (defmethod port-args :phantom [driver]
@@ -2018,7 +2125,7 @@
   Arguments:
 
   - `type` is a keyword determines what driver to use. The supported
-  browsers are `:firefox`, `:chrome`, `:phantom` and `:safari`.
+  browsers are `:firefox`, `:chrome`, `:headless`, `:phantom` and `:safari`.
 
   - `opt` is a map with additional options for a driver. The supported
   options are: